fix: move plugins to single files
